Extract ensureOptionsDefaults
