--- addons/storyshots/storyshots-core/src/api/getIntegrityOptions.js
@@ -1,19 +1,0 @@
-const ignore = ['**/node_modules/**'];
-
-export default function getIntegrityOptions(options) {
-  const { integrityOptions } = options;
-
-  if (integrityOptions === false) {
-    return false;
-  }
-
-  if (typeof integrityOptions !== 'object') {
-    return false;
-  }
-
-  return {
-    ...integrityOptions,
-    ignore: [...ignore, ...(integrityOptions.ignore || [])],
-    absolute: true,
-  };
-}
